requirements.txt
dominate>=2.6.0
imageio>=2.6.1
lmdb>=1.0.0
lpips>=0.1.3
numpy>=1.18.1
opencv-python>=4.4.0.46
pandas>=1.0.1
Pillow>=6.1.0
scikit-image>=0.16.2
scipy>=1.4.1
sk-video>=1.1.10
tensorflow-gpu>=1.15.0
torchvision>=0.7.0
tqdm>=4.42.1
wandb>=0.10.9
clean-fid>=0.1.13
jittor
